Première version de la structure pour le blog, accessible de puis localhost:port/blog et localhost:port/[id]
