+++ pages/blog/[id].js
@@ -1,41 +1,25 @@
-import Link from "next/link";
+import { useRouter } from 'next/router';
+const getPost = async (id) => {
+  const res = await fetch(`https://api.limonecat.com/wp-json/wp/v2/posts/${id}`);
+  return res.json();
+}
 
-export default function Post({ post }) {
+const Post = (props) => {
+  const router = useRouter();
+  const { id } = router.query;
+  const { title } = props.post;
+
   return (
-    <>
-      <main>
-        <Link href="/page3">Page 3</Link>
-        <br />
-        <Link href="/page4">Page 4</Link>
-        <br />
-        <Link href="/page5">Page 5</Link>
-        <br />
-        <h1>{post.title}</h1>
-        <p>{post.body}</p>
-      </main>
-    </>
+      <div>
+        <h1>{title.rendered}</h1>
+      </div>
   );
 }
-// Récupérer les données en fonction de l'id de l'article que je souhaite
-export async function getStaticProps({ params }) {
-  const post = await fetch(
-    `https://jsonplaceholder.typicode.com/posts/${params.id}`
-  ).then((r) => r.json());
-  return {
-    props: {
-      post,
-    },
-  };
+
+Post.getInitialProps = async (ctx) => {
+  const { id } = ctx.query;
+  const post = await getPost(id);
+  return { post };
 }
 
-export async function getStaticPaths() {
-  const posts = await fetch(
-    `https://jsonplaceholder.typicode.com/posts?_limit=5`
-  ).then((r) => r.json());
-  return {
-    paths: posts.map((post) => ({
-      params: { id: post.id.toString() },
-    })),
-    fallback: false,
-  };
-}
+export default Post;
